+++ permuta/patterns/patt.py
@@ -11,11 +11,11 @@ class Patt(abc.ABC):
 
     def avoided_by(self, *patts: "Patt") -> bool:
         """Check if self is avoided by all the provided patterns."""
-        return all(self not in patt for patt in patts)
+        return all(not patt.contains(self) for patt in patts)
 
     def contained_in(self, *patts: "Patt") -> bool:
         """Check if self is a pattern of all the provided patterns."""
-        return all(self in patt for patt in patts)
+        return all(patt.contains(self) for patt in patts)
 
     def count_occurrences_in(self, patt: "Patt") -> int:
         """Count the number of occurrences of self in the pattern."""
@@ -36,5 +36,10 @@ class Patt(abc.ABC):
         """Get the permutation part of the pattern"""
 
     @abc.abstractmethod
-    def __contains__(self, patt: object) -> bool:
+    def _contains(self, patt: "Patt") -> bool:
         """Does pattern contains another?"""
+
+    contains = _contains
+
+    def __contains__(self, patt: "Patt") -> bool:
+        return self._contains(patt)